requirements.txt
certifi==2020.6.20   
chardet==3.0.4       
click==7.1.2
fastapi==0.58.1      
h11==0.9.0
httptools==0.1.1     
idna==2.10
methodtools==0.1.3   
orjson==3.1.2        
pydantic==1.5.1      
python-dotenv==0.13.0
requests==2.24.0 
six==1.15.0      
starlette==0.13.4
urllib3==1.25.9  
uvicorn==0.11.5  
uvloop==0.14.0   
websockets==8.1  
wirerope==0.3.1